+++ src/pages/Signup.tsx
@@ -1,57 +1,9 @@
 import Input from '@/components/common/Input'
 import S from './Signup.module.css'
-import { useState } from 'react'
+import { useSignup } from '@/hooks/auth/useSignup'
 
 export function Signup() {
-  const [birthDate, setBirthDate] = useState({
-    year: '',
-    month: '',
-    day: ''
-  })
-  const [phone, setPhone] = useState('')
-
-  const handleBirthChange = (
-    field: 'year' | 'month' | 'day',
-    value: string
-  ) => {
-    if (value && !/^\d+$/.test(value)) return
-
-    if (field === 'year' && value.length > 4) return
-    if (field === 'month' && (parseInt(value) > 12 || value.length > 2)) return
-    if (field === 'day' && (parseInt(value) > 31 || value.length > 2)) return
-
-    setBirthDate(prev => ({
-      ...prev,
-      [field]: value
-    }))
-  }
-
-  const formatPhoneNumber = (value: string) => {
-    const numbers = value.replace(/[^\d]/g, '')
-
-    // 길이에 따라 하이픈 추가
-    if (numbers.length <= 3) {
-      return numbers
-    } else if (numbers.length <= 7) {
-      return numbers.slice(0, 3) + '-' + numbers.slice(3)
-    } else {
-      return (
-        numbers.slice(0, 3) +
-        '-' +
-        numbers.slice(3, 7) +
-        '-' +
-        numbers.slice(7, 11)
-      )
-    }
-  }
-
-  const handlePhoneChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const value = e.target.value
-    const formattedValue = formatPhoneNumber(value)
-    if (formattedValue.length <= 13) {
-      setPhone(formattedValue)
-    }
-  }
+  const { birthDate, phone, handleBirthChange, handlePhoneChange } = useSignup()
 
   return (
     <div className={S['signup']}>
